Add a new bot, BotMaster
BotMaster manages a list of possible repliers (or "sub-bots") to direct
messages and @reply messages. Upon receiving a message, the available
repliers are scanned until one is found that is willing/able to generate
a reply.

Only one replier, AliveDirectMessageReplier, is currently implemented.
It responds "yes" to direct messages ending in "are you alive?".